+++ tests/column_test.py
@@ -73,3 +73,21 @@ def test_dtype_object_with_arrays():
     df = vaex.from_arrays(z=z)
     assert df.z.tolist()[0].tolist() == x.tolist()
     assert df.z.tolist()[1].tolist() == y.tolist()
+
+
+def test_column_count():
+    x = np.array([1, 2, np.nan])
+    df = vaex.from_arrays(x=x)
+
+    df['new_x'] = df.x + 1
+
+    assert df.column_count() == 2
+
+    # Overwriting a column will rename a column to a hidden column
+    df['new_x'] = df['new_x'].fillna(value=0)
+
+    # By default we do not count hidden columns
+    assert df.column_count() == 2
+
+    # We can count hidden columns if explicity specified
+    assert df.column_count(hidden=True) == 3