+++ tests/values_test.py
@@ -1,5 +1,5 @@
-import pytest
 from common import *
+import pytest
 
 
 def test_values(ds_local):
@@ -10,13 +10,22 @@ def test_values(ds_local):
     assert ds['obj'].values.tolist() == ds.evaluate('obj').tolist()
     assert ds[['x', 'y']].values.tolist() == np.array([ds.evaluate('x'), ds.evaluate('y')]).T.tolist()
     assert ds[['x', 'y']].values.shape == (len(ds), 2)
-    assert ds[['m']].values[9][0] == 77777.0
-    assert ds[['m','x']].values[9][0] == 77777
+    assert ds[['m']].values[9].tolist()[0] == None
+    assert ds[['m', 'x']].values[9].tolist()[0] == None
     # The missing values are included. This may not be the correct behaviour
     assert ds[['x', 'y', 'nm']].values.tolist(), np.array([ds.evaluate('x'), ds.evaluate('y'), ds.evaluate('nm')]).T.tolist()
+
 
 @pytest.mark.skip(reason='TOFIX: obj is now recognized as str')
 def test_object_column_values(ds_local):
     ds = ds_local
     with pytest.raises(ValueError):
         ds[['x', 'name', 'nm', 'obj']].values
+
+
+def test_values_masked():
+    x = np.ma.MaskedArray(data=[1, 2, 3], mask=[False, False, True])
+    y = [10, 20, 30]
+    df = vaex.from_arrays(x=x, y=y)
+
+    assert df[['x', 'y']].values.tolist() ==  [[1., 10.], [2., 20.], [None, 30.]]